Async image loading for AttributedString builder
This simply allows the ImageMarkDownItem to invoke a callback when the image is
loaded (from the cache or network) which will trigger another view configuration.

Due to the processing order of the AttributedString builder I did not see a
simple alternative that would have allowed use of the Kingfisher NSTextAttachment
extension. That is, the builder constructs the attachement as part of composing
the full NSAttributedString for all markdown items prior to creating the parent
UITextView.